pyproject.toml
[project]
name = "openbachelorc"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "frida (==17.6.1)",
    "prompt-toolkit (>=3.0.52,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "openbachelorc", from = "src/launcher" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.10"
pytest = "^9.0.2"
poethepoet = "0.37.0"
pycryptodome = "^3.23.0"

[tool.poetry.scripts]
main = "openbachelorc.main:main"

[tool.poe.tasks]
locker = { script = "locker:main" }
